+++ lib/view/components/raccoon_summary_header.dart
@@ -16,8 +16,6 @@ class RaccoonSummaryHeader extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final colorScheme = theme.colorScheme;
     final statusCode = call.response?.status ?? -1;
 
     return Container(